Operate quickfix or location list window intelligently.
qfutil worked only with quickfix or location list according to the current
qfutil's mode.  So it didn't operate a quickfix window in the location list
mode, even if a quickfix window is opened in the current buffer.

qfutil now tries to operate a quickfix window if it exists in the current
buffer, when qfutil's mode is the location list and vice versa.